refactor: rename component directories
